Don't call our results a "fail"
Failure is a tough word to define in a scientific setting. And it also
paints our work in an unnecessarily bad light. Instead, we should focus
on the drawbacks of the first simplified model (no curvature or solute
effects) and then compare to a more complex model (curvature + solute
effects).